Address PR comments
